Add files via upload
I've created an html page with 3 text boxes in which the user will input 3 numbers :a,b,c
I've added a button that, on click, will calculate ((a*b) + (b*c)) / (a+b). The result will be displayed in an alert.
I've added 3 text boxes in which the user will input 3 numbers :a,b,c
and then a button that, on click, will calculate the maximum number of the three. The result will be displayed in an alert. The program should not crash if the user does not input one or two numbers. 
Cases: 
-If no number is introduced then a message is displayed asking the user to input at least on number.
-If only one number of the three is introduced, that number is the maximum number. 
-If two numbers are introduced then it is displayed the maximum of the two.
-If three numbers are introduced then it will be displayed the maximum of the three.

I also added 2 text boxes in which the user will input 2 numbers: n and m
And then a button that, on click, will calculate the sum of all even numbers between n and m, regardless id n < m or m <  n. The result will be displayed in an alert.